+++ RELEASES.md
@@ -3,6 +3,9 @@
 - Add `get_pos_g_ints` and `get_pos_g_primes` functions
   to return iterators of Gaussian integers and primes with positive real parts.
 - Add `mod-arith` example demonstrating properties of modular arithmetic.
+- Implement [`FromStr`] for `GaussianInt`.
+
+[`FromStr`]: https://doc.rust-lang.org/std/str/trait.FromStr.html
 
 ## v0.6.0 (2022-01-31)